devel_scripts:  Fix issues in Docker image
The `linuxcnc-ethercat` packages had a typo in the sources, causing a
dynamic linker error when the `lcec` component was loaded.  This is
fixed in the cloudsmith repos.  `ENV PACKAGE_UPDATE_SERIAL` is bumped
to pick up that change.

The Docker image build will fail if the EtherCAT master is running.
Print out a big fat warning and exit if it is running.