Fix HAL pins with `-nan` value causing robot startup to fail
* Fix HAL pins with `-nan` value causing robot startup to fail

Fixes ROB-922

The `latency.icomp` has an output pin, `curr_periodf`, that is
initialized with a zero value.  That's connected to the other pins
through the `curr_periodf` signal.  Our recent patch to Machinekit
propagates the value of an output pin to the signal when linked, so
that explains the behavior change.

The `pll` comp was initializing the `periodf` pin to 0.001, so the fix
here is to intialize the `latency` comp `curr_periodf` pin to 0.001
instead.

(Cherry-picked from 6eda3cf2 in ROS 1 stack)